requirements.txt
torch==1.11.0
torchvision==0.12.0
numpy
tqdm
tensorboard
matplotlib
pandas

speechbrain~=0.5.12
datasets
torch_audiomentations
editdistance
wandb
pyctcdecode
torchaudio~=0.11.0
pillow

hydra-core==1.1.1
SoundFile==0.10.3.post1
torchmetrics==1.0.3
